refactor: major code structure change
