+++ src/screens/home/SearchResultItem.tsx
